fix: consistent dropdowns, refactor: ui components
