+++ apps/app/components/issues/select/assignee.tsx
@@ -1,18 +1,17 @@
 import { useState, FC, Fragment } from "react";
 
-import Image from "next/image";
 import { useRouter } from "next/router";
 
 import useSWR from "swr";
 
 // headless ui
 import { Transition, Combobox } from "@headlessui/react";
+// services
+import projectServices from "services/project.service";
+// ui
+import { Avatar } from "components/ui";
 // icons
 import { UserIcon } from "@heroicons/react/24/outline";
-// service
-import projectServices from "services/project.service";
-// types
-import type { IProjectMember } from "types";
 // fetch keys
 import { PROJECT_MEMBERS } from "constants/fetch-keys";
 
@@ -20,35 +19,6 @@ export type IssueAssigneeSelectProps = {
   projectId: string;
   value: string[];
   onChange: (value: string[]) => void;
-};
-
-type AssigneeAvatarProps = {
-  user: IProjectMember | undefined;
-};
-
-export const AssigneeAvatar: FC<AssigneeAvatarProps> = ({ user }) => {
-  if (!user) return <></>;
-
-  if (user.member.avatar && user.member.avatar !== "") {
-    return (
-      <div className="relative h-4 w-4">
-        <Image
-          src={user.member.avatar}
-          alt="avatar"
-          className="rounded-full"
-          layout="fill"
-          objectFit="cover"
-        />
-      </div>
-    );
-  } else
-    return (
-      <div className="grid h-4 w-4 flex-shrink-0 place-items-center rounded-full bg-gray-700 capitalize text-white">
-        {user.member.first_name && user.member.first_name !== ""
-          ? user.member.first_name.charAt(0)
-          : user.member.email.charAt(0)}
-      </div>
-    );
 };
 
 export const IssueAssigneeSelect: FC<IssueAssigneeSelectProps> = ({
@@ -136,14 +106,14 @@ export const IssueAssigneeSelect: FC<IssueAssigneeSelectProps> = ({
                         className={({ active, selected }) =>
                           `${active ? "bg-indigo-50" : ""} ${
                             selected ? "bg-indigo-50 font-medium" : ""
-                          } flex cursor-pointer select-none items-center gap-2 truncate p-2 text-gray-900`
+                          } flex cursor-pointer select-none items-center gap-2 truncate px-2 py-1 text-gray-900`
                         }
                         value={option.value}
                       >
                         {people && (
                           <>
-                            <AssigneeAvatar
-                              user={people?.find((p) => p.member.id === option.value)}
+                            <Avatar
+                              user={people?.find((p) => p.member.id === option.value)?.member}
                             />
                             {option.display}
                           </>
