+++ apps/app/components/ui/custom-listbox.tsx
@@ -2,8 +2,19 @@ import React from "react";
 
 // headless ui
 import { Listbox, Transition } from "@headlessui/react";
-// types
-import { Props } from "./types";
+
+type Props = {
+  title?: string;
+  label?: string;
+  options?: Array<{ display: string; value: any; color?: string; icon?: JSX.Element }>;
+  icon?: JSX.Element;
+  value: any;
+  onChange: (value: any) => void;
+  multiple?: boolean;
+  optionsFontsize?: "sm" | "md" | "lg" | "xl" | "2xl";
+  className?: string;
+  footerOption?: JSX.Element;
+};
 
 export const CustomListbox: React.FC<Props> = ({
   title = "",
