+++ apps/app/components/ui/index.ts
@@ -1,15 +1,17 @@
+export * from "./input";
+export * from "./text-area";
+export * from "./avatar";
 export * from "./button";
 export * from "./custom-listbox";
 export * from "./custom-menu";
 export * from "./custom-select";
+export * from "./datepicker";
 export * from "./empty-space";
 export * from "./header-button";
-export * from "./input";
 export * from "./loader";
+export * from "./multi-input";
 export * from "./outline-button";
+export * from "./progress-bar";
 export * from "./select";
 export * from "./spinner";
-export * from "./text-area";
-export * from "./avatar";
-export * from "./datepicker";
 export * from "./tooltip";
